refactor: reorganize homepage components into dedicated folder
- Move homepage components to (homepage)/components
- Rename HomeExploreSectionCard to ExploreSectionCard
