+++ src/app/(homepage)/components/FeaturedEcosystemProjects.tsx
