+++ src/app/grants/page.tsx
@@ -15,8 +15,8 @@ import { BadgeCardGrid } from '@/components/BadgeCardGrid'
 import { CardGrid } from '@/components/CardGrid'
 import { CardWithBadge } from '@/components/CardWithBadge'
 import { CTASection } from '@/components/CTASection'
+import { ExploreSectionCard } from '@/components/ExploreSectionCard'
 import { FeaturedGrantsGraduates } from '@/components/FeaturedGrantGraduates'
-import { HomeExploreSectionCard } from '@/components/HomeExploreSectionCard'
 import { PageHeader } from '@/components/PageHeader'
 import { PageLayout } from '@/components/PageLayout'
 import { PageSection } from '@/components/PageSection'
@@ -82,7 +82,7 @@ export default function Grants() {
             const { title, description, icon, cta } = card
 
             return (
-              <HomeExploreSectionCard
+              <ExploreSectionCard
                 key={title}
                 cta={cta}
                 heading={{
@@ -95,7 +95,7 @@ export default function Grants() {
                 }}
               >
                 {description}
-              </HomeExploreSectionCard>
+              </ExploreSectionCard>
             )
           })}
         </CardGrid>
@@ -136,7 +136,7 @@ export default function Grants() {
             const { title, description, icon } = data
 
             return (
-              <HomeExploreSectionCard
+              <ExploreSectionCard
                 key={title}
                 heading={{
                   tag: 'h3',
@@ -148,7 +148,7 @@ export default function Grants() {
                 }}
               >
                 {description}
-              </HomeExploreSectionCard>
+              </ExploreSectionCard>
             )
           })}
         </CardGrid>
